updated the pendant, make a damageable pendant and make other pendant not take damage from combat
