+++ src/main/java/teamport/aether/item/accessory/pendant/ItemCombatPendant.java
@@ -1,0 +1,33 @@
+package teamport.aether.item.accessory.pendant;
+
+import net.minecraft.core.item.IArmorItem;
+import net.minecraft.core.item.material.ArmorMaterial;
+import org.jspecify.annotations.Nullable;
+
+public class ItemCombatPendant extends ItemPendant implements IArmorItem {
+
+
+    public ItemCombatPendant(String translationKey, String namespaceId, int id, ArmorMaterial material) {
+        super(translationKey, namespaceId, id, material);
+    }
+
+    @Override
+    public float getArmorPieceProtectionPercentage() {
+        return this.armorPieceProtection() / 20.0F;
+    }
+
+    @Override
+    public @Nullable ArmorMaterial getArmorMaterial() {
+        return this.material;
+    }
+
+
+    @Override
+    public int armorPieceProtection() {
+        return 1;
+    }
+
+    @Override
+    public int getArmorPiece() {
+        return -1;
+    }}
